Start breaking up functionality
